third_party/blink/web_tests/svg/dom/SVGRect-expected.txt

This test checks the SVGRect API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial rect values
PASS rect.x is 0
PASS rect.y is 0
PASS rect.width is 0
PASS rect.height is 0


Check assigning rects
PASS rect.x = 100 is 100
PASS rect.y = 200 is 200
PASS rect.width = 300 is 300
PASS rect.height = 400 is 400
Check that the rect contains the correct values
PASS rect.x is 100
PASS rect.y is 200
PASS rect.width is 300
PASS rect.height is 400


Check assigning invalid rects
PASS rect.x = rect threw exception TypeError: Failed to set the 'x' property on 'SVGRect': The provided float value is non-finite..
PASS rect.y = null is null
PASS rect.width = 'aString' threw exception TypeError: Failed to set the 'width' property on 'SVGRect': The provided float value is non-finite..
PASS rect.height = svgElement threw exception TypeError: Failed to set the 'height' property on 'SVGRect': The provided float value is non-finite..


Check that the rect contains the correct values
PASS rect.x is 100
PASS rect.y is 0
PASS rect.width is 300
PASS rect.height is 400
PASS successfullyParsed is true

TEST COMPLETE
